claude-code/cli-generator/pyproject.toml
[project]
name = "cli-generator"
version = "0.1.0"
description = "Generate CLI tools from natural language descriptions"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pydantic>=2.0",
    "pydantic-ai>=0.1",
    "jinja2>=3.1",
    "httpx>=0.25",
    "python-dotenv>=1.0",
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
[project.scripts]
cli-gen = "cli_generator.cli:main"
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
